requirements.txt
python-dotenv
Flask
pytest
pytest-flask
firebase-admin
jsonschema
flask-cors
numpy
PyEphem